Trying to write run loop for Java properly
This may not be a good solution yet, but the number of failing tests
decreases with 15, so it's going in the right direction at least.

The problem with the previous implementation was that it broke out of
the loop if a timer stopped event machine, while the C implementations
still ran the remaining next tick code.

It's actually probably more complicated than that though, since some of
the code did have a next_tick that should run before the timers ran
which stopped the EM loop.